Add files via upload
This code is provided as an example in our academic paper titled "An Efficient MATLAB Code for Stress-constrained Level Set Topology Optimization Method". This code is intended for educational purposes only. If you use this code in your research or work, we kindly ask you to cite our paper as the source. Please note that the authors and the university take no responsibility for the misuse of this code or any consequences thereof. Use it responsibly and with due consideration to the principles of academic integrity.